+++ test/in/subCode.js
@@ -1,0 +1,1 @@
+export const doNotLogThisFunctionName = function () { }