+++ test/out/subCode.js
@@ -1,0 +1,10 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.doNotLogThisFunctionName = void 0;
+
+var doNotLogThisFunctionName = function doNotLogThisFunctionName() {};
+
+exports.doNotLogThisFunctionName = doNotLogThisFunctionName;