+++ test/transform.js
@@ -1,8 +1,36 @@
 const babel = require('@babel/core');
 const fs = require('fs');
 const basePath = process.cwd();
-const decodefilePath = basePath + '/test/source.js';
-const encodefilePath = basePath + '/test/transformed.js'
-const fileValue = fs.readFileSync(decodefilePath);
-const enCode = babel.transform(fileValue);
-fs.writeFileSync(encodefilePath, enCode.code);
+
+const inFiles = basePath + '/test/in/';
+const directoryList = [inFiles];
+
+let inFilesPath = [];
+let outFilesPath = [];
+
+while (directoryList.length) {
+  let path = directoryList[0];
+  let filse = fs.readdirSync(path);
+  directoryList.shift();
+  filse.forEach((item) => {
+    if (fs.lstatSync(path + item).isDirectory() === true) {
+      directoryList.push(path + item + '/');
+      let outDirPath = basePath + '/test/out/' + path.substring(inFiles.length, path.length) + item;
+      if (!fs.existsSync(outDirPath)) {
+        fs.mkdirSync(outDirPath)
+      }
+    } else {
+      inFilesPath.push(path + item);
+    }
+  })
+}
+
+inFilesPath.forEach((item) => {
+  outFilesPath.push(basePath + '/test/out/' + item.substring(inFiles.length, item.length))
+})
+
+for (let i = 0; i < inFilesPath.length; i++) {
+  let outCode = babel.transformFileSync(inFilesPath[i]);
+  fs.writeFileSync(outFilesPath[i], outCode.code);
+}
+